+++ extensions/iceworks-doctor/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.1.1
 
 - docs: add logo
+- feat: Add welcome page
 
 ## 0.1.0